refactor: implement full DI for EntityStore and extract EntityMutationService
- Created EntityMutationService to handle all state-changing operations.
- Updated EntityStore to accept services via Dependency Injection.
- Maintained backward compatibility for tests.
- Reduced EntityStore from 920 to 450 lines.
- Updated God File Analysis report.
